base pass

+++ Assets/Scripts/Games/Ball/BallState/BallStateData.cs
@@ -3,14 +3,14 @@
 
 public class BallStateData
 {
-    public int LockDuration { get; set; }
+    public float LockDuration { get; set; }
     
     public static BallStateData Build()
     {
         return new BallStateData();
     }
     
-    public BallStateData SetLockDuration(int duration)
+    public BallStateData SetLockDuration(float duration)
     {
         LockDuration = duration;
         return this;
